+++ client/.eslintrc.js
@@ -16,7 +16,6 @@ module.exports = {
     'import/extensions': ['.js', '.jsx', '.ts', '.tsx'],
   },
   rules: {
-    eqeqeq: ['warn'],
     'import/order': [
       'warn',
       {