+++ client/src/containers/analysis-visualization/analysis-filters/more-filters/component.tsx
@@ -53,7 +53,6 @@ interface ApiTreeResponse {
 }
 
 const DEFAULT_QUERY_OPTIONS = {
-  staleTime: 2 * 60 * 1000, // 2 minutes
   select: (data: ApiTreeResponse[]) => {
     const sorted = recursiveSort(data, 'name');
     return sorted.map((item) => recursiveMap(item, ({ id, name }) => ({ label: name, value: id })));
@@ -72,7 +71,6 @@ const MoreFilters = () => {
     [materials, origins, suppliers, locationTypes],
   );
 
-  // Initial state from redux
   const [selectedFilters, setSelectedFilters] = useState(moreFilters);
 
   const materialIds = useMemo(
@@ -95,17 +93,12 @@ const MoreFilters = () => {
     [selectedFilters.locationTypes],
   );
 
-  const [counter, setCounter] = useState<number>(0);
+  const [counter, setCounter] = useState(0);
 
   // Only the changes are applied when the user clicks on Apply
   const handleApply = useCallback(() => {
     dispatch(setFilters(selectedFilters));
   }, [dispatch, selectedFilters]);
-
-  // Close filters window
-  const handleCancel = useCallback(() => {
-    setSelectedFilters(moreFilters);
-  }, [moreFilters]);
 
   // Restoring state from initial state only internally,
   // the user have to apply the changes
@@ -122,16 +115,8 @@ const MoreFilters = () => {
   );
 
   useEffect(() => {
-    const counters = Object.values(moreFilters).map((value) => value.length);
-    const total = counters.reduce((a, b) => a + b);
-    setCounter(total);
+    setSelectedFilters(moreFilters);
   }, [moreFilters]);
-
-  // Check difference between current selection of filters and filters already applied
-  const hasChangesToApply = useMemo(
-    () => JSON.stringify(selectedFilters) !== JSON.stringify(moreFilters),
-    [selectedFilters, moreFilters],
-  );
 
   const { reference, floating, strategy, x, y, context } = useFloating({
     // open: isOpen,
@@ -196,10 +181,23 @@ const MoreFilters = () => {
   );
 
   const { data: locationTypeOptions } = useLocationTypes(
-    {},
     {
-      ...DEFAULT_QUERY_OPTIONS,
-      select: undefined,
+      materialIds,
+      originIds,
+      supplierIds,
+      scenarioIds,
+    },
+    {
+      onSuccess: (_locationTypeOptions) => {
+        // * every time new location types are fetched, we need to validate if the previous location types selected are still
+        // * available in the new options. Otherwise, we will remove them from the current selection.
+        setSelectedFilters((filters) => ({
+          ...filters,
+          locationTypes: _locationTypeOptions.filter(({ value }) =>
+            locationTypesIds.includes(value),
+          ),
+        }));
+      },
     },
   );
 
@@ -228,6 +226,12 @@ const MoreFilters = () => {
     reviewFilterContent('origins', origins, origins);
     reviewFilterContent('suppliers', suppliers, suppliers);
   }, [locationTypes, materialOptions, materials, origins, reviewFilterContent, suppliers]);
+
+  useEffect(() => {
+    const counters = Object.values(moreFilters).map((value) => value.length);
+    const total = counters.reduce((a, b) => a + b);
+    setCounter(total);
+  }, [moreFilters]);
 
   useEffect(() => {
     handleScenarioChange();
@@ -336,7 +340,6 @@ const MoreFilters = () => {
                     className="px-9"
                     onClick={() => {
                       close();
-                      handleCancel();
                     }}
                   >
                     Cancel
@@ -344,8 +347,11 @@ const MoreFilters = () => {
                   <Button
                     variant="primary"
                     className="flex-grow"
-                    onClick={handleApply}
-                    disabled={!hasChangesToApply}
+                    onClick={() => {
+                      handleApply();
+                      close();
+                    }}
+                    data-testid="more-filters-apply-btn"
                   >
                     Apply
                   </Button>